+++ src/components/Flashcard/Flashcard.css
@@ -1,14 +1,12 @@
 .FlashcardContainer {
   height: 100%; }
-  .FlashcardContainer .mobileSwipeContainer {
-    height: 80%; }
-  .FlashcardContainer .Flashcard {
-    height: 80%; }
   .FlashcardContainer .notecard {
-    background: -webkit-gradient(linear, left top, left bottom, color-stop(15%, #ffffff), color-stop(16%, #F94545), color-stop(16%, #ffffff), color-stop(25%, #ffffff), color-stop(26%, #85b2d3), color-stop(26%, #ffffff), color-stop(35%, #ffffff), color-stop(36%, #85b2d3), color-stop(36%, #ffffff), color-stop(45%, #ffffff), color-stop(46%, #85b2d3), color-stop(46%, #ffffff), color-stop(55%, #ffffff), color-stop(56%, #85b2d3), color-stop(56%, #ffffff), color-stop(65%, #ffffff), color-stop(66%, #85b2d3), color-stop(66%, #ffffff), color-stop(75%, #ffffff), color-stop(76%, #85b2d3), color-stop(76%, #ffffff), color-stop(85%, #ffffff), color-stop(86%, #85b2d3), color-stop(86%, #ffffff), color-stop(95%, #ffffff), color-stop(96%, #85b2d3), color-stop(96%, #ffffff));
-    background: -webkit-linear-gradient(top, #ffffff 15%, #F94545 16%, #ffffff 16%, #ffffff 25%, #85b2d3 26%, #ffffff 26%, #ffffff 35%, #85b2d3 36%, #ffffff 36%, #ffffff 45%, #85b2d3 46%, #ffffff 46%, #ffffff 55%, #85b2d3 56%, #ffffff 56%, #ffffff 65%, #85b2d3 66%, #ffffff 66%, #ffffff 75%, #85b2d3 76%, #ffffff 76%, #ffffff 85%, #85b2d3 86%, #ffffff 86%, #ffffff 95%, #85b2d3 96%, #ffffff 96%);
-    background: -o-linear-gradient(top, #ffffff 15%, #F94545 16%, #ffffff 16%, #ffffff 25%, #85b2d3 26%, #ffffff 26%, #ffffff 35%, #85b2d3 36%, #ffffff 36%, #ffffff 45%, #85b2d3 46%, #ffffff 46%, #ffffff 55%, #85b2d3 56%, #ffffff 56%, #ffffff 65%, #85b2d3 66%, #ffffff 66%, #ffffff 75%, #85b2d3 76%, #ffffff 76%, #ffffff 85%, #85b2d3 86%, #ffffff 86%, #ffffff 95%, #85b2d3 96%, #ffffff 96%);
-    background: linear-gradient(to bottom, #ffffff 15%, #F94545 16%, #ffffff 16%, #ffffff 25%, #85b2d3 26%, #ffffff 26%, #ffffff 35%, #85b2d3 36%, #ffffff 36%, #ffffff 45%, #85b2d3 46%, #ffffff 46%, #ffffff 55%, #85b2d3 56%, #ffffff 56%, #ffffff 65%, #85b2d3 66%, #ffffff 66%, #ffffff 75%, #85b2d3 76%, #ffffff 76%, #ffffff 85%, #85b2d3 86%, #ffffff 86%, #ffffff 95%, #85b2d3 96%, #ffffff 96%); }
+    background: #D3CCE3;
+    /* fallback for old browsers */
+    background: -webkit-linear-gradient(to left, #E9E4F0, #D3CCE3);
+    /* Chrome 10-25, Safari 5.1-6 */
+    background: linear-gradient(to left, #E9E4F0, #D3CCE3);
+    /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */ }
   .FlashcardContainer .customBgImg {
     background-repeat: no-repeat;
     background-position: center;
@@ -32,6 +30,7 @@
         .FlashcardContainer .customBgImg.Bulgakov .lead, .FlashcardContainer .customBgImg.Freud .lead, .FlashcardContainer .customBgImg.Tolstoy .lead {
           font-size: 1.1em; } }
   .FlashcardContainer .Flashcard {
+    height: 80%;
     cursor: pointer;
     border-radius: 2px;
     font-size: 1.5em;
@@ -45,6 +44,8 @@
     .FlashcardContainer .Flashcard.newCard {
       transition: 0s; }
     .FlashcardContainer .Flashcard .lead {
+      position: absolute;
+      top: 50%;
       font-size: 2.3em;
       text-align: center;
       left: 0;
