+++ src/components/Flashcard/Flashcard.scss
@@ -1,20 +1,18 @@
 .FlashcardContainer {
   height: 100%;
 
-  // child div isn't in markup - part of ReactSwipe dependency (this allows for swiping on the entire card)
-  .mobileSwipeContainer {
-    height: 80%;
-  }
-
-  .Flashcard {
-    height: 80%;
-  }
-
   .notecard {
-    background: -webkit-gradient(linear, left top, left bottom, color-stop(15%, #ffffff),color-stop(16%, #F94545),color-stop(16%, #ffffff),color-stop(25%, #ffffff),color-stop(26%, #85b2d3),color-stop(26%, #ffffff), color-stop(35%, #ffffff),color-stop(36%, #85b2d3),color-stop(36%, #ffffff),color-stop(45%, #ffffff),color-stop(46%, #85b2d3),color-stop(46%, #ffffff), color-stop(55%, #ffffff),color-stop(56%, #85b2d3),color-stop(56%, #ffffff), color-stop(65%, #ffffff),color-stop(66%, #85b2d3),color-stop(66%, #ffffff), color-stop(75%, #ffffff),color-stop(76%, #85b2d3),color-stop(76%, #ffffff), color-stop(85%, #ffffff),color-stop(86%, #85b2d3),color-stop(86%, #ffffff), color-stop(95%, #ffffff),color-stop(96%, #85b2d3),color-stop(96%, #ffffff));
-    background: -webkit-linear-gradient(top, #ffffff 15%,#F94545 16%,#ffffff 16%,#ffffff 25%,#85b2d3 26%,#ffffff 26%, #ffffff 35%,#85b2d3 36%,#ffffff 36%,#ffffff 45%,#85b2d3 46%,#ffffff 46%, #ffffff 55%,#85b2d3 56%,#ffffff 56%, #ffffff 65%,#85b2d3 66%,#ffffff 66%, #ffffff 75%,#85b2d3 76%,#ffffff 76%, #ffffff 85%,#85b2d3 86%,#ffffff 86%, #ffffff 95%,#85b2d3 96%,#ffffff 96%);
-    background: -o-linear-gradient(top, #ffffff 15%,#F94545 16%,#ffffff 16%,#ffffff 25%,#85b2d3 26%,#ffffff 26%, #ffffff 35%,#85b2d3 36%,#ffffff 36%,#ffffff 45%,#85b2d3 46%,#ffffff 46%, #ffffff 55%,#85b2d3 56%,#ffffff 56%, #ffffff 65%,#85b2d3 66%,#ffffff 66%, #ffffff 75%,#85b2d3 76%,#ffffff 76%, #ffffff 85%,#85b2d3 86%,#ffffff 86%, #ffffff 95%,#85b2d3 96%,#ffffff 96%);
-    background: linear-gradient(to bottom, #ffffff 15%,#F94545 16%,#ffffff 16%,#ffffff 25%,#85b2d3 26%,#ffffff 26%, #ffffff 35%,#85b2d3 36%,#ffffff 36%,#ffffff 45%,#85b2d3 46%,#ffffff 46%, #ffffff 55%,#85b2d3 56%,#ffffff 56%, #ffffff 65%,#85b2d3 66%,#ffffff 66%, #ffffff 75%,#85b2d3 76%,#ffffff 76%, #ffffff 85%,#85b2d3 86%,#ffffff 86%, #ffffff 95%,#85b2d3 96%,#ffffff 96%);
+    // background: -webkit-gradient(linear, left top, left bottom, color-stop(15%, #ffffff),color-stop(16%, #F94545),color-stop(16%, #ffffff),color-stop(25%, #ffffff),color-stop(26%, #85b2d3),color-stop(26%, #ffffff), color-stop(35%, #ffffff),color-stop(36%, #85b2d3),color-stop(36%, #ffffff),color-stop(45%, #ffffff),color-stop(46%, #85b2d3),color-stop(46%, #ffffff), color-stop(55%, #ffffff),color-stop(56%, #85b2d3),color-stop(56%, #ffffff), color-stop(65%, #ffffff),color-stop(66%, #85b2d3),color-stop(66%, #ffffff), color-stop(75%, #ffffff),color-stop(76%, #85b2d3),color-stop(76%, #ffffff), color-stop(85%, #ffffff),color-stop(86%, #85b2d3),color-stop(86%, #ffffff), color-stop(95%, #ffffff),color-stop(96%, #85b2d3),color-stop(96%, #ffffff));
+    // background: -webkit-linear-gradient(top, #ffffff 15%,#F94545 16%,#ffffff 16%,#ffffff 25%,#85b2d3 26%,#ffffff 26%, #ffffff 35%,#85b2d3 36%,#ffffff 36%,#ffffff 45%,#85b2d3 46%,#ffffff 46%, #ffffff 55%,#85b2d3 56%,#ffffff 56%, #ffffff 65%,#85b2d3 66%,#ffffff 66%, #ffffff 75%,#85b2d3 76%,#ffffff 76%, #ffffff 85%,#85b2d3 86%,#ffffff 86%, #ffffff 95%,#85b2d3 96%,#ffffff 96%);
+    // background: -o-linear-gradient(top, #ffffff 15%,#F94545 16%,#ffffff 16%,#ffffff 25%,#85b2d3 26%,#ffffff 26%, #ffffff 35%,#85b2d3 36%,#ffffff 36%,#ffffff 45%,#85b2d3 46%,#ffffff 46%, #ffffff 55%,#85b2d3 56%,#ffffff 56%, #ffffff 65%,#85b2d3 66%,#ffffff 66%, #ffffff 75%,#85b2d3 76%,#ffffff 76%, #ffffff 85%,#85b2d3 86%,#ffffff 86%, #ffffff 95%,#85b2d3 96%,#ffffff 96%);
+    // background: linear-gradient(to bottom, #ffffff 15%,#F94545 16%,#ffffff 16%,#ffffff 25%,#85b2d3 26%,#ffffff 26%, #ffffff 35%,#85b2d3 36%,#ffffff 36%,#ffffff 45%,#85b2d3 46%,#ffffff 46%, #ffffff 55%,#85b2d3 56%,#ffffff 56%, #ffffff 65%,#85b2d3 66%,#ffffff 66%, #ffffff 75%,#85b2d3 76%,#ffffff 76%, #ffffff 85%,#85b2d3 86%,#ffffff 86%, #ffffff 95%,#85b2d3 96%,#ffffff 96%);
+    //
+    background: #D3CCE3;
+    /* fallback for old browsers */
+    background: -webkit-linear-gradient(to left, #E9E4F0, #D3CCE3);
+    /* Chrome 10-25, Safari 5.1-6 */
+    background: linear-gradient(to left, #E9E4F0, #D3CCE3);
+    /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
   }
 
   .customBgImg {
@@ -39,6 +37,7 @@
     &.Tolstoy {
       background-color: rgba(0,0,0,0.7);
       color: #fff;
+
       .lead {
         @media(max-width: 880px) {
           font-size: 1.7em;
@@ -52,41 +51,42 @@
       }
     }
   }
-    .Flashcard {
-      cursor: pointer;
-      border-radius: 2px;
-      font-size: 1.5em;
-      -webkit-box-shadow: 0 1px 7px rgba(0, 0, 0, 0.05);
-      box-shadow: 0 1px 7px rgba(0, 0, 0, 0.05);
-      -webkit-transition: 0.6s;
-      -o-transition: 0.6s;
-      transition: 0.6s;
 
-      &:hover {
-        border-radius: 3px;
-      }
+  .Flashcard {
+    height: 80%;
+    cursor: pointer;
+    border-radius: 2px;
+    font-size: 1.5em;
+    -webkit-box-shadow: 0 1px 7px rgba(0, 0, 0, 0.05);
+    box-shadow: 0 1px 7px rgba(0, 0, 0, 0.05);
+    -webkit-transition: 0.6s;
+    -o-transition: 0.6s;
+    transition: 0.6s;
 
-      &.newCard {
-        transition: 0s;
-      }
-
-      .lead {
-        font-size: 2.3em;
-        text-align: center;
-        // position: absolute;
-        left:0;
-        right:0;
-        margin: 0 auto;
-        @media(max-width: 770px) {
-          font-size: 2em;
-        }
-        @media(max-width: 500px) {
-          font-size: 1.6em;
-        }
-      }
+    &:hover {
+      border-radius: 3px;
     }
 
+    &.newCard {
+      transition: 0s;
+    }
 
+    .lead {
+      position: absolute;
+      top: 50%;
+      font-size: 2.3em;
+      text-align: center;
+      left: 0;
+      right: 0;
+      margin: 0 auto;
+      @media(max-width: 770px) {
+        font-size: 2em;
+      }
+      @media(max-width: 500px) {
+        font-size: 1.6em;
+      }
+    }
+  }
   // // animates card on click - it's a little annoying though so I've turned it off
   // .flipped {
   //   -webkit-transform: rotateY(180deg);
@@ -97,7 +97,4 @@
   //     transform: rotateY(180deg);
   //   }
   // }
-
-
-
 }
