+++ src/components/Flashcard/FlashcardButtons.scss
@@ -1,6 +1,7 @@
 .button-toolbar {
   height: 15%;
   width: 100%;
+  margin-left: 0;
   button {
     width: percentage(1/3);
     height: 100%;
@@ -8,5 +9,10 @@
     border-radius: 0;
     border: 0;
     margin-left: 0 !important; /* bootstrap override */
+  }
+  &.mobile-button-toolbar {
+    button {
+      width: 100%;
+    }
   }
 }
