+++ src/components/Flashcard/FlashcardMobileDevice.scss
@@ -1,0 +1,12 @@
+.FlashcardMobileDevice {
+  height: 80%;
+  .mobileSwipeContainer {
+    height: 100%;
+    div {
+      height: 100%;
+      .Flashcard {
+        height: 100%;
+      }
+    }
+  }
+}
