+++ src/components/Navigation/Navigation.css
@@ -4,3 +4,11 @@
   height: 5vh;
   margin-bottom: 0;
   display: block; }
+  .navbar-inverse .navbar-collapse.collapse.in {
+    position: fixed;
+    z-index: 1;
+    width: 100%; }
+    .navbar-inverse .navbar-collapse.collapse.in .navbar-nav {
+      margin-top: 0; }
+  .navbar-inverse li {
+    background-color: #333; }
